Move parser::value::PlaceHolder to tables::
